made description component a server comp
